add template to bit

+++ Math/nCk.cpp
@@ -3,9 +3,8 @@ using namespace std;
 
 typedef long long ll;
 
-const long long N = 2e5 + 1, mod = 1e9 + 7;
-long long fn[N], fd[N], iv[N];
-
+const ll N = 2e5 + 1, mod = 1e9 + 7;
+ll fn[N], fd[N], iv[N];
 
 void initnck() {
     iv[1] = fn[1] = fd[1] = fn[0] = fd[0] = 1;
